@@ -10,6 +10,14 @@ import Foundation
 
 class Router {
 
+    private static let dateFormatter = {
+        let newDateFormatter = DateFormatter()
+        newDateFormatter.dateFormat = "EEE, dd MMM yyyy HH:mm:ssZ"
+        newDateFormatter.locale = Locale(identifier: "en_US_POSIX")
+        newDateFormatter.timeZone = TimeZone(abbreviation: "UTC")
+        return newDateFormatter
+    }()
+
     class func make(request: Request) async throws(APIError) -> APIResult {
         var components = URLComponents()
         components.scheme = "https"
@@ -30,6 +38,8 @@ class Router {
             urlRequest.setValue(header.value, forHTTPHeaderField: header.key)
         }
 
+        logger.info(urlRequest.debugDescription)
+
         do {
             let result = try await URLSession.shared.data(for: urlRequest)
             guard
@@ -39,10 +49,17 @@ class Router {
             }
 
             if 200..<300 ~= httpResult.statusCode {
+                var serverDate: Date?
+                if let responseHeaders = httpResult.allHeaderFields as? [String: String],
+                   let serverDateHeader = responseHeaders["Date"] {
+                    serverDate = dateFormatter.date(from: serverDateHeader)
+                }
                 return .init(
-                    data: result.0
+                    data: result.0,
+                    dateOfResponse: serverDate
                 )
             } else {
+                logger.err(String(data: result.0, encoding: .utf8) ?? "null")
                 throw APIError.server(httpResult.statusCode, result.0)
             }
         } catch {
@@ -54,6 +71,7 @@ class Router {
 struct APIResult {
 
     let data: Data
+    let dateOfResponse: Date?
 }
 
 public enum APIError: Error {
