@@ -13,6 +13,23 @@ import ComposableArchitecture
 public extension Reducer {
 
     func preloadResources(
+        for snippet: TWSSnippet,
+        using api: APIDependency
+    ) async -> [TWSSnippet.Attachment: String] {
+        let attachments = snippet.dynamicResources ?? []
+        guard !attachments.isEmpty else { return [:] }
+
+        return await _preloadResources(
+            homepages: [.init(
+                url: snippet.target,
+                contentType: .html
+            )],
+            resources: attachments,
+            using: api
+        )
+    }
+
+    func preloadResources(
         for project: TWSProject,
         using api: APIDependency
     ) async -> [TWSSnippet.Attachment: String] {
@@ -20,8 +37,15 @@ public extension Reducer {
             .compactMap(\.dynamicResources)
             .flatMap { $0 }
 
-        guard !attachments.isEmpty else { return [:] }
-        return await _preloadResources(resources: attachments, using: api)
+        return await _preloadResources(
+            homepages: project.snippets.map {
+                TWSSnippet.Attachment(
+                    url: $0.target,
+                    contentType: .html)
+            },
+            resources: attachments,
+            using: api
+        )
     }
 
     func preloadResources(
@@ -30,10 +54,18 @@ public extension Reducer {
     ) async -> [TWSSnippet.Attachment: String] {
         let attachments = sharedSnippet.snippet.dynamicResources ?? []
         guard !attachments.isEmpty else { return [:] }
-        return await _preloadResources(resources: attachments, using: api)
+        return await _preloadResources(
+            homepages: [.init(
+                url: sharedSnippet.snippet.target,
+                contentType: .html
+            )],
+            resources: attachments,
+            using: api
+        )
     }
 
     private func _preloadResources(
+        homepages: [TWSSnippet.Attachment],
         resources: [TWSSnippet.Attachment],
         using api: APIDependency
     ) async -> [TWSSnippet.Attachment: String] {
@@ -41,11 +73,13 @@ public extension Reducer {
             of: (TWSSnippet.Attachment, String)?.self,
             returning: [TWSSnippet.Attachment: String].self
         ) { [api] group in
-            for resource in resources {
+            // Use a set to download each resource only once, even if it is used in multiple snippets
+            for resource in Set(homepages + resources) {
                 group.addTask { [resource] in
                     do {
                         let payload = try await api.getResource(resource)
-                        return (resource, payload)
+                        if !payload.isEmpty { return (resource, payload) }
+                        return nil
                     } catch {
                         return nil
                     }
