@@ -2,7 +2,6 @@ import Foundation
 import ComposableArchitecture
 import TWSSettings
 import TWSSnippets
-import TWSModels
 import TWSUniversalLinks
 
 @Reducer
@@ -31,7 +30,7 @@ public struct TWSCoreFeature {
     public enum Action {
         case settings(TWSSettingsFeature.Action)
         case snippets(TWSSnippetsFeature.Action)
-        case snippetsDidChange([TWSSnippet])
+        case snippetsDidChange
         case universalLinks(TWSUniversalLinksFeature.Action)
     }
 
