@@ -8,36 +8,49 @@
 
 import SwiftUI
 import TWSKit
+import TWSUI
 
 @MainActor
 struct ProjectView: View {
 
     @State private var viewModel: ProjectViewModel
-    @State private var selectedID: UUID
+    @State private var selectedID: TWSSnippet.ID
 
-    init(viewModel: ProjectViewModel, selectedID: UUID) {
+    init(viewModel: ProjectViewModel, selectedID: TWSSnippet.ID) {
         _viewModel = .init(initialValue: viewModel)
         _selectedID = .init(initialValue: selectedID)
     }
 
     var body: some View {
         TabView(selection: $selectedID) {
-            ForEach(viewModel.snippets) { snippet in
+            ForEach(viewModel.tabSnippets) { snippet in
                 ProjectSnippetView(
                     snippet: snippet,
-                    manager: viewModel.manager
+                    organizationID: "\(viewModel.manager.id.hashValue)"
                 )
-                .tabItem { Text("\(snippet.id.uuidString.suffix(4))") }
+                .tabItem {
+                    if let tabName = snippet.props?[.tabName, as: \.string] {
+                        Text(tabName)
+                    }
+
+                    if let icon = snippet.props?[.tabIcon, as: \.string] {
+                        Image(systemName: icon)
+                    }
+                }
                 .tag(snippet.id)
             }
         }
         .ignoresSafeArea(edges: .bottom)
         .task {
-            await viewModel.start()
             await viewModel.startupInitTasks()
         }
         .sheet(item: $viewModel.universalLinkLoadedProject) {
             ProjectView(viewModel: $0.viewModel, selectedID: $0.selectedID)
+                .twsEnable(using: viewModel.universalLinkLoadedProject!.viewModel.manager)
         }
+        .fullScreenCover(isPresented: $viewModel.presentPopups, content: {
+            TWSPopupView(isPresented: $viewModel.presentPopups, manager: viewModel.manager)
+                .twsEnable(using: viewModel.manager)
+        })
     }
 }
