@@ -15,13 +15,25 @@ class ProjectViewModel {
 
     private let _id = UUID().uuidString.suffix(4)
     let manager: TWSManager
-    private(set) var snippets: [TWSSnippet]
+    private(set) var tabSnippets: [TWSSnippet]
+    private(set) var popupSnippets: [TWSSnippet]
     let destinationID = UUID()
     var universalLinkLoadedProject: LoadedProjectInfo?
+    var presentPopups: Bool = false
+
+    private var clearedPopupSnippets: [TWSSnippet] = []
 
     init(manager: TWSManager) {
-        self.snippets = manager.snippets
+        let snippets = manager.snippets
+        self.tabSnippets = snippets.filter(\.isTab)
+        self.popupSnippets = snippets.filter { _ in
+            // As of now, because `type` needs to be removed with TWS-212,
+            // We have no way to detect pop-ups (Check TWSPopUpViewModel too)
+            // Before: return snippet.type == .popup
+            return false
+        }
         self.manager = manager
+        self.presentPopups = !popupSnippets.isEmpty
         // Do not call `.run()` in the initializer! SwiftUI views can recreate multiple instances of the same view.
         // Therefore, the initializer should be free of any business logic.
         // Calling `run` here will trigger a refresh, potentially causing excessive updates.
@@ -30,10 +42,6 @@ class ProjectViewModel {
 
     deinit {
         print("DEINIT ->", _id, "ProjectViewModel")
-    }
-
-    func start() async {
-        manager.run()
     }
 
     func startupInitTasks() async {
@@ -51,9 +59,17 @@ class ProjectViewModel {
                     selectedID: project.snippet.id
                 )
 
-            case let .snippetsUpdated(updatedSnippets):
+            case .snippetsUpdated:
                 print("->", _id, "Received event: snippets updated")
-                self.snippets = updatedSnippets
+                let snippets = manager.snippets
+                self.tabSnippets = snippets.filter(\.isTab)
+                self.popupSnippets = snippets.filter({ _ in
+                    // As of now, because `type` needs to be removed with TWS-212,
+                    // We have no way to detect pop-ups (Check TWSPopUpViewModel too)
+                    // Before: snippet.type == .popup && self.canShowPopupSnippet(snippet)
+                    return false
+                })
+                self.presentPopups = !self.popupSnippets.isEmpty
 
             @unknown default:
                 break
@@ -61,7 +77,16 @@ class ProjectViewModel {
         }
 
         print("->", _id, "Stopped listening")
-        snippets = []
+        tabSnippets = []
+        popupSnippets = []
         universalLinkLoadedProject = nil
+    }
+
+    public func addClearedPopup(_ snippet: TWSSnippet) {
+        self.clearedPopupSnippets.append(snippet)
+    }
+
+    public func canShowPopupSnippet(_ snippet: TWSSnippet) -> Bool {
+        return !clearedPopupSnippets.contains(snippet)
     }
 }
