@@ -13,16 +13,30 @@ import TWSKit
 @Observable
 class TWSViewModel {
 
-    let manager = TWSFactory.new(with: .init(
-        organizationID: "357b24f6-4714-42b1-9e15-0b07cae2fcd6",
-        projectID: "4166c981-56ae-4007-bc93-28875e6a2ca5"
-    ))
+    let manager = TWSFactory.new(
+        with: .init(
+            organizationID: "inova.tws",
+            projectID: "4166c981-56ae-4007-bc93-28875e6a2ca5"
+        )
+    )
     let destinationID = UUID()
-    var snippets: [TWSSnippet]
+    var tabSnippets: [TWSSnippet]
+    var popupSnippets: [TWSSnippet]
     var universalLinkLoadedProject: LoadedProjectInfo?
+    var presentPopups: Bool = false
+
+    private var clearedPopupSnippets: [TWSSnippet] = []
 
     init() {
-        snippets = manager.snippets
+        let snippets = manager.snippets
+        self.tabSnippets = snippets.filter(\.isTab)
+        self.popupSnippets = snippets.filter { _ in
+            // As of now, because `type` needs to be removed with TWS-212,
+            // We have no way to detect pop-ups (Check TWSPopUpViewModel too)
+            // Before: return snippet.type == .popup
+            return false
+        }
+        self.presentPopups = !popupSnippets.isEmpty
         // Do not call `.run()` in the initializer! SwiftUI views can recreate multiple instances of the same view.
         // Therefore, the initializer should be free of any business logic.
         // Calling `run` here will trigger a refresh, potentially causing excessive updates.
@@ -30,10 +44,6 @@ class TWSViewModel {
 
     func handleIncomingUrl(_ url: URL) {
         manager.handleIncomingUrl(url)
-    }
-
-    func start() async {
-        manager.run()
     }
 
     func startupInitTasks() async {
@@ -49,12 +59,30 @@ class TWSViewModel {
                     selectedID: project.snippet.id
                 )
 
-            case .snippetsUpdated(let snippets):
-                self.snippets = snippets
+            case .snippetsUpdated:
+                let snippets = manager.snippets
+                self.tabSnippets = snippets.filter(\.isTab)
+                self.popupSnippets = snippets.filter({ _ in
+                    // As of now, because `type` needs to be removed with TWS-212,
+                    // We have no way to detect pop-ups (Check TWSPopUpViewModel too)
+                    // Before: return snippet.type == .popup
+                    //    && self.canShowPopupSnippet(snippet)
+
+                    return false
+                })
+                self.presentPopups = !self.popupSnippets.isEmpty
 
             default:
                 assertionFailure("Unhandled stream event")
             }
         }
+    }
+
+    public func addClearedPopup(_ snippet: TWSSnippet) {
+        self.clearedPopupSnippets.append(snippet)
+    }
+
+    public func canShowPopupSnippet(_ snippet: TWSSnippet) -> Bool {
+        return !clearedPopupSnippets.contains(snippet)
     }
 }
