@@ -5,9 +5,12 @@ import Combine
 internal import TWSCore
 internal import ComposableArchitecture
 internal import TWSLogger
+internal import TWSSnippets
+internal import TWSSnippet
 
 /// A class that handles all the communication between your app and the SDK's functionalities
 @MainActor
+@Observable
 public final class TWSManager: Identifiable {
 
     let observer: AnyPublisher<TWSStreamEvent, Never>
@@ -18,6 +21,8 @@ public final class TWSManager: Identifiable {
 
     private let initDate: Date
     private let _id = UUID().uuidString.suffix(4)
+    private var _stateSync: AnyCancellable?
+    private var isSetup = false
 
     init(
         store: StoreOf<TWSCoreFeature>,
@@ -30,8 +35,10 @@ public final class TWSManager: Identifiable {
         self.initDate = Date()
         self.snippetHeightProvider = SnippetHeightProviderImpl()
         self.navigationProvider = NavigationProviderImpl()
+        self.snippets = store.snippets.snippets.elements.map(\.snippet)
 
         logger.info("INIT TWSManager \(_id)")
+        _syncState()
     }
 
     deinit {
@@ -39,17 +46,30 @@ public final class TWSManager: Identifiable {
         MainActor.assumeIsolated { TWSFactory.destroy(configuration: configuration) }
     }
 
+    private func _syncState() {
+        _stateSync = observer
+            .compactMap {
+                switch $0 {
+                case .snippetsUpdated: return ()
+                case .universalLinkSnippetLoaded: return nil
+                }
+            }
+            .sink(receiveValue: { [weak self] _ in
+                guard let weakSelf = self else { return }
+                weakSelf.snippets = weakSelf.store.snippets.snippets.elements.map(\.snippet)
+            })
+    }
+
     // MARK: - Public
 
     /// A getter for the list of loaded snippets
-    public var snippets: [TWSSnippet] {
-        precondition(Thread.isMainThread, "`snippets()` can only be called on main thread")
-        return store.snippets.snippets.elements.map(\.snippet)
-    }
+    public internal(set) var snippets: [TWSSnippet]
 
     /// A function that starts loading snippets and listen for changes
     public func run() {
         precondition(Thread.isMainThread, "`run()` can only be called on main thread")
+        defer { isSetup = true }
+        guard !isSetup else { return }
         store.send(.snippets(.business(.load)))
     }
 
@@ -98,6 +118,15 @@ public final class TWSManager: Identifiable {
         store.send(.snippets(.business(.set(source: source))))
     }
 
+    /// Defines a custom set of local properties that will be injected into the ``TWSView``.
+    /// - Parameters:
+    ///   - localProps: A dictionary containing the local properties to inject.
+    ///   - id: The identifier of the snippet to associate with this dictionary.
+    public func set(localProps: [String: TWSSnippet.Props], to id: TWSSnippet.ID) {
+        precondition(Thread.isMainThread, "`set(customProps:,to:)` can only be called on main thread")
+        store.send(.snippets(.business(.setLocalProps(props: (id, localProps)))))
+    }
+
     /// An async function that gathers all the logs in the current session
     /// - Parameter reportFiltering: A function that lets you define which data of the logs do you want displayed
     /// - Returns: An URL of the file that has all the logs written inside
@@ -134,8 +163,21 @@ public final class TWSManager: Identifiable {
         else {
             return
         }
-        store.send(.snippets(.business(
-            .snippets(.element(id: snippet.id, action: .business(.update(height: height, forId: displayID))))
-        )))
+
+        let snippetAction: TWSSnippetFeature.Action = .business(.update(
+            height: height,
+            forId: displayID
+        ))
+
+        let snippetsAction: TWSSnippetsFeature.Action = .business(
+            .snippets(
+                .element(
+                    id: snippet.id,
+                    action: snippetAction
+                )
+            )
+        )
+
+        store.send(.snippets(snippetsAction))
     }
 }
