@@ -15,6 +15,9 @@ extension TWSSnippetsFeature {
     // This is because TCA uses combine pipelines, which means the same hashable value
     // could potentially leak to all active stores.
     enum CancelID: Hashable {
-        case socket(TWSConfiguration), reconnect(TWSConfiguration)
+        case socket(TWSConfiguration),
+             reconnect(TWSConfiguration),
+             showSnippet(TWSSnippet.ID),
+             hideSnippet(TWSSnippet.ID)
     }
 }
