@@ -31,9 +31,9 @@ final class SnippetsTests: XCTestCase {
 
     @MainActor
     func testAPIShouldNotOverrideState() async {
-        let s1ID = UUID()
-        let s2ID = UUID()
-        let s3ID = UUID()
+        let s1ID = "1"
+        let s2ID = "2"
+        let s3ID = "3"
 
         let snippets: [TWSSnippet] = [
             .init(id: s1ID, target: URL(string: "https://www.google.com")!),
@@ -47,7 +47,8 @@ final class SnippetsTests: XCTestCase {
             initialState: state,
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: snippets)}
+                $0.api.getProject = { [socketURL] _ in (TWSProject(listenOn: socketURL, snippets: snippets), nil)}
+                $0.api.getResource = { _ in return "" }
             }
         )
 
@@ -72,9 +73,9 @@ final class SnippetsTests: XCTestCase {
 
     @MainActor
     func testAPIShouldRemoveFromStateOnceNotReturned() async {
-        let s1ID = UUID()
-        let s2ID = UUID()
-        let s3ID = UUID()
+        let s1ID = "1"
+        let s2ID = "2"
+        let s3ID = "3"
 
         let snippets: [TWSSnippet] = [
             .init(id: s1ID, target: URL(string: "https://www.google.com")!),
@@ -86,7 +87,8 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: snippets)}
+                $0.api.getProject = { [socketURL] _ in (TWSProject(listenOn: socketURL, snippets: snippets), nil)}
+                $0.api.getResource = { _ in return "" }
             }
         )
 
@@ -100,8 +102,8 @@ final class SnippetsTests: XCTestCase {
 
         // Send for the second time without one element. Snippet should be removed from state
 
-        store.dependencies.api.getProject = { _ in
-            TWSProject(listenOn: self.socketURL, snippets: [snippets[1], snippets[2]])
+        store.dependencies.api.getProject = { [socketURL] _ in
+            (TWSProject(listenOn: socketURL, snippets: [snippets[1], snippets[2]]), nil)
         }
 
         await store.send(.business(.load)).finish()
@@ -112,9 +114,9 @@ final class SnippetsTests: XCTestCase {
 
     @MainActor
     func testAPIShouldAddNewWhenReturned() async {
-        let s1ID = UUID()
-        let s2ID = UUID()
-        let s3ID = UUID()
+        let s1ID = "1"
+        let s2ID = "2"
+        let s3ID = "3"
 
         let snippets: [TWSSnippet] = [
             .init(id: s1ID, target: URL(string: "https://www.google.com")!),
@@ -126,9 +128,10 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in
-                    TWSProject(listenOn: self.socketURL, snippets: [snippets[0], snippets[2]])
+                $0.api.getProject = { [socketURL] _ in
+                    (TWSProject(listenOn: socketURL, snippets: [snippets[0], snippets[2]]), nil)
                 }
+                $0.api.getResource = { _ in return "" }
             }
         )
 
@@ -141,7 +144,9 @@ final class SnippetsTests: XCTestCase {
         }
 
         // Send for the second time with new element. Snippet should be added in right order
-        store.dependencies.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: snippets)}
+        store.dependencies.api.getProject = { [socketURL] _ in
+            (TWSProject(listenOn: socketURL, snippets: snippets), nil)
+        }
 
         await store.send(.business(.load)).finish()
         await store.receive(\.business.projectLoaded.success) {
@@ -151,9 +156,9 @@ final class SnippetsTests: XCTestCase {
 
     @MainActor
     func testAPIOrderChange() async {
-        let s1ID = UUID()
-        let s2ID = UUID()
-        let s3ID = UUID()
+        let s1ID = "1"
+        let s2ID = "2"
+        let s3ID = "3"
 
         let snippets: [TWSSnippet] = [
             .init(id: s1ID, target: URL(string: "https://www.google.com")!),
@@ -167,7 +172,8 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: snippets)}
+                $0.api.getProject = { [socketURL] _ in (TWSProject(listenOn: socketURL, snippets: snippets), nil)}
+                $0.api.getResource = { _ in return "" }
             }
         )
 
@@ -181,8 +187,8 @@ final class SnippetsTests: XCTestCase {
 
         // Send response for the second time but change the order
 
-        store.dependencies.api.getProject = { _ in
-            TWSProject(listenOn: self.socketURL, snippets: [snippets[1], snippets[2], snippets[0]])
+        store.dependencies.api.getProject = { [socketURL] _ in
+            (TWSProject(listenOn: socketURL, snippets: [snippets[1], snippets[2], snippets[0]]), nil)
         }
 
         await store.send(.business(.load)).finish()
@@ -193,10 +199,10 @@ final class SnippetsTests: XCTestCase {
 
     @MainActor
     func testAddingAndRemoving() async {
-        let s1ID = UUID()
-        let s2ID = UUID()
-        let s3ID = UUID()
-        let s4ID = UUID()
+        let s1ID = "1"
+        let s2ID = "2"
+        let s3ID = "3"
+        let s4ID = "4"
 
         let snippets: [TWSSnippet] = [
             .init(id: s1ID, target: URL(string: "https://www.google.com")!),
@@ -211,9 +217,10 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in
-                    TWSProject(listenOn: self.socketURL, snippets: [snippets[0], snippets[1], snippets[2]])
+                $0.api.getProject = { [socketURL] _ in
+                    (TWSProject(listenOn: socketURL, snippets: [snippets[0], snippets[1], snippets[2]]), nil)
                 }
+                $0.api.getResource = { _ in return "" }
             }
         )
 
@@ -223,12 +230,13 @@ final class SnippetsTests: XCTestCase {
         await store.receive(\.business.projectLoaded.success) {
             $0.snippets = .init(uniqueElements: [snippetsStates[0], snippetsStates[1], snippetsStates[2]])
             $0.socketURL = self.socketURL
+            $0.snippetDates = [:]
         }
 
         // Send response for the second time but remove some and add some
 
-        store.dependencies.api.getProject = { _ in
-            TWSProject(listenOn: self.socketURL, snippets: [snippets[0], snippets[2], snippets[3]])
+        store.dependencies.api.getProject = { [socketURL] _ in
+            (TWSProject(listenOn: socketURL, snippets: [snippets[0], snippets[2], snippets[3]]), nil)
         }
 
         await store.send(.business(.load)).finish()
@@ -244,7 +252,7 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsObserverFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in .init(listenOn: self.socketURL, snippets: [])}
+                $0.api.getProject = { [socketURL] _ in (.init(listenOn: socketURL, snippets: []), nil)}
                 $0.socket.get = { _, _ in .init() }
                 $0.socket.connect = { _ in .makeStream().stream }
                 $0.socket.closeConnection = { _ in }
@@ -275,7 +283,7 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: [])}
+                $0.api.getProject = { [socketURL] _ in (TWSProject(listenOn: socketURL, snippets: []), nil)}
             }
         )
 
