@@ -44,12 +44,17 @@ final class ResourcesAggregationTests: XCTestCase {
             )
         ]
 
-        let preloadedAttachments = Dictionary(
+        var preloadedAttachments = Dictionary(
             uniqueKeysWithValues:
                 attachments.map {
                     ($0, $0.url.absoluteString)
                 }
         )
+
+        preloadedAttachments[.init(
+            url: url,
+            contentType: .html
+        )] = url.absoluteString
 
         let sharedSnippet = TWSSharedSnippet(
             organization: .init(
@@ -59,18 +64,19 @@ final class ResourcesAggregationTests: XCTestCase {
                 id: "00000000-0000-0000-0000-000000000001"
             ),
             snippet: .init(
-                id: UUID(uuidString: "00000000-0000-0000-0000-000000000002")!,
+                id: "00000000-0000-0000-0000-000000000002",
                 target: url,
-                dynamicResources: attachments
+                dynamicResources: attachments,
+                status: "enabled"
             )
         )
 
-        let store = TestStore(
+        let store = await TestStore(
             initialState: .init(),
             reducer: { TWSUniversalLinksFeature() }
             ,
             withDependencies: {
-                $0.api.getSnippetBySharedId = { _, _ in return sharedSnippet }
+                $0.api.getSnippetBySharedId = { @Sendable _, _ in return sharedSnippet }
                 $0.api.getResource = { attachment in return attachment.url.absoluteString }
             }
         )
